refactor: Simplify serverless architecture and use @rollup/plugin-typescript
- Replaced rollup-plugin-typescript2 with @rollup/plugin-typescript
- Removed redundant serverless build configurations
- Kept only the lightweight serverless version (244KB minified)
- Simplified file naming: removed '.lite' suffix from serverless builds
- Updated documentation to reflect simplified architecture
- Fixed TypeScript plugin configuration for proper declaration file generation
