+++ package.json
@@ -35,8 +35,7 @@
   "scripts": {
     "build": "rm -rf lib && rollup -c --bundleConfigAsCjs",
     "build:serverless": "rollup -c rollup.config.serverless.js --bundleConfigAsCjs",
-    "build:serverless:lite": "rollup -c rollup.config.serverless.lite.js --bundleConfigAsCjs",
-    "build:all": "yarn build && yarn build:serverless:lite",
+    "build:all": "yarn build && yarn build:serverless",
     "preparemetadata": "rm -rf resources && node .scripts/prepare.js && rm -rf resources/libphonenumber",
     "prepublishOnly": "yarn build:all",
     "format": "biome format --write .",
@@ -57,6 +56,7 @@
     "@rollup/plugin-json": "^6.1.0",
     "@rollup/plugin-node-resolve": "^16.0.1",
     "@rollup/plugin-terser": "^0.4.4",
+    "@rollup/plugin-typescript": "^12.1.4",
     "@types/bson": "^4.2.4",
     "@types/jest": "^30.0.0",
     "@types/node": "^24.3.0",
@@ -64,7 +64,6 @@
     "jest": "^30.1.1",
     "prettier": "^3.6.2",
     "rollup": "^4.49.0",
-    "rollup-plugin-typescript2": "^0.36.0",
     "shelljs": "^0.10.0",
     "ts-jest": "^29.4.1",
     "tslib": "^2.8.1",
