+++ rollup.config.js
@@ -1,4 +1,4 @@
-import typescript from 'rollup-plugin-typescript2'
+import typescript from '@rollup/plugin-typescript'
 import pkg from './package.json'
 
 export default {
@@ -14,5 +14,8 @@ export default {
     },
   ],
   external: [...Object.keys(pkg.dependencies || {}), ...Object.keys(pkg.peerDependencies || {})],
-  plugins: [typescript()],
+  plugins: [typescript({
+    tsconfig: './tsconfig.json',
+    declarationDir: './lib',
+  })],
 }
